refactor search logic out of ui class
